Improve code quality
